1、实现新的UI(未完成)

+++ videoplayer/src/main/java/com/yoy/videoPlayer/processing/PlayStateListener.kt
@@ -15,5 +15,5 @@ interface PlayStateListener {
      * 更新播放时间
      * @param time 时间,单位ms
      */
-    fun onPlayTime(time:Int)
+    fun onPlayTime(time:Long)
 }
